+++ Venmo/src/test/java/com/braintreepayments/api/VenmoClientUnitTest.java
@@ -157,7 +157,7 @@ public class VenmoClientUnitTest {
         ArgumentCaptor<Intent> captor = ArgumentCaptor.forClass(Intent.class);
 
         verify(activity).startActivity(captor.capture());
-        verify(braintreeClient).sendAnalyticsEvent("android.pay-with-venmo.app-store.invoked");
+        verify(braintreeClient).sendAnalyticsEvent("android.pay-with-venmo.app-store.invoked", null);
     }
 
     @Test
@@ -220,7 +220,45 @@ public class VenmoClientUnitTest {
         assertEquals("session-id", intent.getSessionId());
 
         verify(sharedPrefsWriter).persistVenmoVaultOption(activity, false);
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", "venmo-payment-context-id");
+    }
+
+    @Test
+    public void tokenizeVenmoAccount_whenCreatePaymentContextSucceeds_withObserver_withEmptyPaymentContextId_launchesObserverWithVenmoIntentData_andSendsAnalytics() {
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .configuration(venmoEnabledConfiguration)
+                .sessionId("session-id")
+                .integration("custom")
+                .authorizationSuccess(clientToken)
+                .build();
+
+        VenmoApi venmoApi = new MockVenmoApiBuilder()
+                .createPaymentContextSuccess("")
+                .build();
+
+        when(deviceInspector.isVenmoAppSwitchAvailable(activity)).thenReturn(true);
+
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
+        request.setProfileId("sample-venmo-merchant");
+        request.setShouldVault(false);
+
+        VenmoClient sut = new VenmoClient(activity, lifecycle, braintreeClient, venmoApi, sharedPrefsWriter, deviceInspector);
+        VenmoLifecycleObserver observer = mock(VenmoLifecycleObserver.class);
+        sut.observer = observer;
+        sut.tokenizeVenmoAccount(activity, request);
+
+        ArgumentCaptor<VenmoIntentData> captor = ArgumentCaptor.forClass(VenmoIntentData.class);
+        verify(observer).launch(captor.capture());
+
+        VenmoIntentData intent = captor.getValue();
+        assertEquals("", intent.getPaymentContextId());
+        assertSame(venmoEnabledConfiguration, intent.getConfiguration());
+        assertEquals("custom", intent.getIntegrationType());
+        assertEquals("sample-venmo-merchant", intent.getProfileId());
+        assertEquals("session-id", intent.getSessionId());
+
+        verify(sharedPrefsWriter).persistVenmoVaultOption(activity, false);
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", null);
     }
 
     @Test
@@ -249,7 +287,7 @@ public class VenmoClientUnitTest {
         sut.tokenizeVenmoAccount(activity, request);
 
         verify(listener).onVenmoFailure(captor.capture());
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
         assertEquals("Cannot collect customer data when ECD is disabled. Enable this feature in the Control Panel to collect this data.", captor.getValue().getMessage());
     }
 
@@ -279,7 +317,36 @@ public class VenmoClientUnitTest {
 
         InOrder inOrder = Mockito.inOrder(activity, braintreeClient);
 
-        inOrder.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started");
+        inOrder.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", "venmo-payment-context-id");
+    }
+
+    @Test
+    public void tokenizeVenmoAccount_whenCreatePaymentContext_withEmptyPaymentContextId_collectAddressWithEcdEnabled() throws JSONException {
+        venmoEnabledConfiguration = Configuration.fromJson(Fixtures.CONFIGURATION_WITH_PAY_WITH_VENMO_ECD_ENABLED);
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .configuration(venmoEnabledConfiguration)
+                .sessionId("session-id")
+                .integration("custom")
+                .authorizationSuccess(clientToken)
+                .build();
+
+        VenmoApi venmoApi = new MockVenmoApiBuilder()
+                .createPaymentContextSuccess("")
+                .build();
+
+        when(deviceInspector.isVenmoAppSwitchAvailable(activity)).thenReturn(true);
+
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
+        request.setProfileId("sample-venmo-merchant");
+        request.setCollectCustomerBillingAddress(true);
+
+        VenmoClient sut = new VenmoClient(null, null, braintreeClient, venmoApi, sharedPrefsWriter, deviceInspector);
+        sut.setListener(listener);
+        sut.tokenizeVenmoAccount(activity, request);
+
+        InOrder inOrder = Mockito.inOrder(activity, braintreeClient);
+
+        inOrder.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", null);
     }
 
     @Test
@@ -309,7 +376,7 @@ public class VenmoClientUnitTest {
         ArgumentCaptor<Intent> captor = ArgumentCaptor.forClass(Intent.class);
         inOrder.verify(activity).startActivityForResult(captor.capture(), eq(BraintreeRequestCodes.VENMO));
 
-        inOrder.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started");
+        inOrder.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", "venmo-payment-context-id");
 
         Intent intent = captor.getValue();
         assertEquals(new ComponentName("com.venmo", "com.venmo.controller.SetupMerchantActivity"), intent.getComponent());
@@ -389,7 +456,7 @@ public class VenmoClientUnitTest {
                 ArgumentCaptor.forClass(Exception.class);
         verify(listener).onVenmoFailure(captor.capture());
         assertEquals("Configuration fetching error", captor.getValue().getMessage());
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
     }
 
     @Test
@@ -410,7 +477,7 @@ public class VenmoClientUnitTest {
                 ArgumentCaptor.forClass(AppSwitchNotAvailableException.class);
         verify(listener).onVenmoFailure(captor.capture());
         assertEquals("Venmo is not enabled", captor.getValue().getMessage());
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
     }
 
     @Test
@@ -434,7 +501,7 @@ public class VenmoClientUnitTest {
                 ArgumentCaptor.forClass(AppSwitchNotAvailableException.class);
         verify(venmoTokenizeAccountCallback).onResult(captor.capture());
         assertEquals("Venmo is not installed", captor.getValue().getMessage());
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
     }
 
     @Test
@@ -624,7 +691,7 @@ public class VenmoClientUnitTest {
         sut.setListener(listener);
         sut.tokenizeVenmoAccount(activity, request);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected", null);
     }
 
     @Test
@@ -648,8 +715,8 @@ public class VenmoClientUnitTest {
         sut.observer = mock(VenmoLifecycleObserver.class);
         sut.tokenizeVenmoAccount(activity, request);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected");
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected", null);
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.started", "venmo-payment-context-id");
     }
 
     @Test
@@ -828,8 +895,8 @@ public class VenmoClientUnitTest {
         verify(listener).onVenmoFailure(captor.capture());
         assertEquals("Venmo is not installed", captor.getValue().getMessage());
 
-        order.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected");
-        order.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        order.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.selected", null);
+        order.verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
     }
 
     @Test
@@ -852,7 +919,7 @@ public class VenmoClientUnitTest {
         sut.tokenizeVenmoAccount(activity, request);
 
         verify(listener).onVenmoFailure(graphQLError);
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failed", null);
     }
 
     @Test
@@ -895,7 +962,7 @@ public class VenmoClientUnitTest {
         assertEquals("fake-venmo-nonce", nonce.getString());
         assertEquals("venmojoe", nonce.getUsername());
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success", null);
     }
 
     @Test
@@ -918,7 +985,7 @@ public class VenmoClientUnitTest {
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
         verify(onActivityResultCallback).onResult(null, graphQLError);
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failure");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failure", null);
     }
 
     @Test
@@ -980,7 +1047,7 @@ public class VenmoClientUnitTest {
 
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success", null);
     }
 
     @Test
@@ -988,7 +1055,7 @@ public class VenmoClientUnitTest {
         VenmoClient sut = new VenmoClient(activity, lifecycle, braintreeClient, venmoApi, sharedPrefsWriter, deviceInspector);
         sut.onActivityResult(activity, AppCompatActivity.RESULT_CANCELED, new Intent(), onActivityResultCallback);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.canceled");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", null);
     }
 
     @Test
@@ -1101,7 +1168,7 @@ public class VenmoClientUnitTest {
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
         verify(onActivityResultCallback).onResult(any(VenmoAccountNonce.class), (Exception) isNull());
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.success"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.success", null);
     }
 
     @Test
@@ -1126,7 +1193,7 @@ public class VenmoClientUnitTest {
         VenmoClient sut = new VenmoClient(activity, lifecycle, braintreeClient, venmoApi, sharedPrefsWriter, deviceInspector);
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.success"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.success", null);
     }
 
     @Test
@@ -1180,7 +1247,7 @@ public class VenmoClientUnitTest {
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
         verify(onActivityResultCallback).onResult(null, error);
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.failed"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.failed", null);
     }
 
     @Test
@@ -1206,7 +1273,7 @@ public class VenmoClientUnitTest {
                 .putExtra(EXTRA_PAYMENT_METHOD_NONCE, "nonce");
         sut.onActivityResult(activity, AppCompatActivity.RESULT_OK, intent, onActivityResultCallback);
 
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.failed"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.failed", null);
     }
 
     @Test
@@ -1308,7 +1375,7 @@ public class VenmoClientUnitTest {
         assertEquals("fake-venmo-nonce", nonce.getString());
         assertEquals("venmojoe", nonce.getUsername());
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success", null);
     }
 
     @Test
@@ -1331,7 +1398,7 @@ public class VenmoClientUnitTest {
         sut.onVenmoResult(venmoResult);
 
         verify(listener).onVenmoFailure(graphQLError);
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failure");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.failure", null);
     }
 
     @Test
@@ -1389,7 +1456,7 @@ public class VenmoClientUnitTest {
         VenmoResult venmoResult = new VenmoResult("payment-context-id", "some-nonce", "venmo-username", null);
         sut.onVenmoResult(venmoResult);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.success", null);
     }
 
     @Test
@@ -1400,7 +1467,7 @@ public class VenmoClientUnitTest {
         VenmoResult venmoResult = new VenmoResult("payment-context-id", null, null, new UserCanceledException("User canceled Venmo."));
         sut.onVenmoResult(venmoResult);
 
-        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.canceled");
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", null);
     }
 
     @Test
@@ -1488,7 +1555,7 @@ public class VenmoClientUnitTest {
         sut.onVenmoResult(venmoResult);
 
         verify(listener).onVenmoSuccess(venmoAccountNonce);
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.success"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.success", null);
     }
 
     @Test
@@ -1517,7 +1584,7 @@ public class VenmoClientUnitTest {
         sut.onVenmoResult(venmoResult);
 
         verify(listener).onVenmoSuccess(venmoAccountNonce);
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.success"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.success", null);
     }
 
     @Test
@@ -1544,7 +1611,7 @@ public class VenmoClientUnitTest {
         sut.onVenmoResult(venmoResult);
 
         verify(listener).onVenmoFailure(error);
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.failed"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.failed", null);
     }
 
     @Test
@@ -1574,7 +1641,7 @@ public class VenmoClientUnitTest {
         sut.onVenmoResult(venmoResult);
 
         verify(listener).onVenmoFailure(error);
-        verify(braintreeClient).sendAnalyticsEvent(endsWith("pay-with-venmo.vault.failed"));
+        verify(braintreeClient).sendAnalyticsEvent("pay-with-venmo.vault.failed", null);
     }
 
     @Test